Move manager to its own file
